+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Implementation changes and bug fixes
 - [PR #2328](https://github.com/rqlite/rqlite/pull/2328): Store CDC database in dedicated directory.
 - [PR #2329](https://github.com/rqlite/rqlite/pull/2329): Switch go mod to version v9.
+- [PR #2330](https://github.com/rqlite/rqlite/pull/2330): Support NONE queries on restart even if node not in cluster.
 
 ## v9.0.0 (September 15th 2025)
 This release introduces Change Data Capture (CDC). When CDC is enabled any changes to the underlying SQLite database are streamed to a HTTP endpoint. Check out the [announcement blog post](https://philipotoole.com/rqlite-9-0-real-time-change-data-capture-for-distributed-sqlite/) for more details.